Add files via upload
My python for the basic financial analysis. It includes data cleaning code as well as functions for a correlation matrix, basic statistics, and pickup/dropoff-specific tip stats. It should be commented pretty well, but a lot of this code is rough and as such I probably wouldn't use it just yet.